Save to store and AsyncStorage not working
Lines 52-55 is where I tried to save the title first to the store then to local storage, followed by the questions.

I had to save the title along with the questions here because the rerender after saving the title only produces errors trying to count the number of questions associated with the title, which is undefined.