Code Refactor for Count
Fix, Refactor

Prior to this commit, my App was constantly crashing because of an array length check function that runs as soon as the app compiles (Decks.js).

I moved away from having to run this function each time by adding a piece of state that keeps track of the cards in  each deck and gets updated by 1 each time a set of question and answer is added.

The App launched correctly now, however the saveNewDeck action is still broken for now.