refactor: DeckView Independence
The DeckaView component is now independent of other components and screens sending all data needed to render a view. All that is needed now is the title key for the deck to be viewed. This refactor also allows for more frequent routes to this DeckView from the addTitle and addQuestion screens; these however will be achived in the next commit.